lerobot_gui/requirements_gui.txt
PySide6>=6.6
PyYAML>=6.0
numpy>=1.24
